Old code import
This is a sanitzed version of the AutoHotKey script I built for use on my work computer. The original version was developed in a private repo and is not suitable for publishing (it contained hard coded work emails and phone numbers).

This should be considered Version 0.5, and will be heavily re-written due to the scope of the project having gotten out of hand.

Signed-off-by: Justin Kenney <JustinKenney45@gmail.com>